+++ src/models/v2/project-metadata/person.ts
@@ -33,10 +33,10 @@ export class Person extends BaseProjectMetadata {
     jobTitle: string[] = [];
 
     @JsonProperty(Constants.DspIsMemberOf, UnionOrganizationIdConverter)
-    memberOf: Organization[] | IId = [];
+    memberOf: Organization[] | IId[] = [];
 
     @JsonProperty(Constants.DspSameAs, UnionUrlStringConverter, true)
-    sameAs?: IUrl = undefined;
+    sameAs?: IUrl[] = undefined;
 
     constructor() {
         super(Constants.DspPerson);